\name{AR1}
\docType{data}
\alias{AR1}
\title{Simlated AR(1) process}
\description{
  artificial datas

 \emph{number of observations} :  100

}
\usage{data(AR1)}
\format{A dataframe containing :
  \describe{
    \item{ID}{identification number}
    \item{x}{simulated}
    \item{y}{simulated}
  }
}
\references{
    Davidson, R.  and  James G.  MacKinnon (2004) \emph{Econometric Theory and Methods}, New York, Oxford University Press, chapter 10.
}
\keyword{datasets}