% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_tokencounts}
\alias{as_tokencounts}
\title{Utility functions to map sparse matrices to token count lists}
\usage{
as_tokencounts(x)
}
\arguments{
\item{x}{a sparse Matrix}
}
\value{
a list with elements tokens, counts and vocab with the content of \code{x}
}
\description{
Convert a sparse matrix to counts of tokens
}
\examples{
library(udpipe)
library(Matrix)
data(brussels_reviews_anno, package = "udpipe")
x  <- subset(brussels_reviews_anno, upos \%in\% "NOUN")
x  <- document_term_frequencies(x, document = "doc_id", term = "lemma")
x  <- document_term_matrix(x)
tk <- as_tokencounts(x)
str(tk)

## Test to do the other way around: tokencounts to sparse matrix
as_dtm <- function(tokens, counts, vocab){
  nm  <- seq_len(length(tokens))
  mat <- sparseMatrix(i = unlist(Map(nm, tokens, 
                                     f = function(nm, key) rep(nm, length(key)))), 
                      j = unlist(tokens, use.names = FALSE), 
                      x = unlist(counts, use.names = FALSE))
  colnames(mat) <- vocab
  mat
}
x_back      <- as_dtm(tokens = tk$tokens, counts = tk$counts, vocab = tk$vocab)
rownames(x) <- NULL
all.equal(x, x_back)
}
