\name{Count}
\docType{data}
\alias{Count}
\title{Number of doctor visits}
\description{
  a cross-section

 \emph{number of observations} :  485

 \emph{observation} :  individuals

}
\usage{data(Count)}
\format{A dataframe containing :
  \describe{
    \item{ID}{the observation number}
    \item{doctor}{the number of doctor visits}
    \item{children}{the number of children in the household}
    \item{access}{is a measure of access to health care}
    \item{health}{a measure of health status (larger positive numbers are  associated with poorer health)}
  }
}
\note{\dQuote{There are 485 observations, all from the AFDC part of the sample.}}
\source{
    Gurmu, Shiferaw (1997) \dQuote{Semiparametric estimation of hurdle regression models with an application to medicaid utilization}, \emph{Journal of Applied Econometrics }, \bold{12(3)}, 225-242.
}
\references{
    Davidson, R.  and  James G.  MacKinnon (2004) \emph{Econometric Theory and Methods}, New York, Oxford University Press, chapter 11.
}
\keyword{datasets}