\name{Demsup}
\docType{data}
\alias{Demsup}
\title{Artificial data for demand-supply model}
\description{
  artificial datas

 \emph{number of observations} :  120

}
\usage{data(Demsup)}
\format{A dataframe containing :
  \describe{
    \item{ID}{the observation number}
    \item{q}{the log of quantity}
    \item{p}{the log of price}
    \item{x2}{a variable that appears only in the demand equation}
    \item{x3}{a variable that appears only in the demand equation}
    \item{x4}{a variable that appears only in the supply equation}
    \item{x5}{a variable that appears only in the supply equation}
  }
}
\references{
    Davidson, R.  and  James G.  MacKinnon (2004) \emph{Econometric Theory and Methods}, New York, Oxford University Press, chapter 8 and 12.
}
\keyword{datasets}