\name{Earnings}
\docType{data}
\alias{Earnings}
\title{Earnings for three age groups}
\description{
  a cross-section from 1988-1989

 \emph{number of observations} :  4266

 \emph{observation} :  individuals

   \emph{country} :  United States

}
\usage{data(Earnings)}
\format{A dataframe containing :
  \describe{
    \item{ID}{the observation number}
    \item{g1}{dummy variable for individuals in group 1}
    \item{g2}{dummy variable for individuals in group 2}
    \item{g3}{dummy variable for individuals in group 3}
    \item{y}{average annual earnings, in 1982 US dollars}
  }
}
\note{\dQuote{They are average annual earnings in 1988 and 1989, in 1982 US dollars, for individuals in three age groups. See the paper for details.}}
\source{
    Mills, Jeffery A.  and  Sourushe  Zandvakili (1997) \dQuote{Statistical Inference via  Bootstrapping for Measures of Inequality}, \emph{Journal of Applied Econometrics }, \bold{12(2)}, pp. 133-150.
}
\references{
    Davidson, R.  and  James G.  MacKinnon (2004) \emph{Econometric Theory and Methods}, New York, Oxford University Press, chapter 5 and 7.
}
\keyword{datasets}