\name{Intratesm}
\docType{data}
\alias{Intratesm}
\title{Interest rates (monthly)}
\description{
  monthly observations from 1955-1 to 2001-12

 \emph{number of observations} :  564

 \emph{observation} :  Country

   \emph{country} :  United States

}
\usage{data(Intratesm)}
\format{A dataframe containing :
  \describe{
    \item{ID}{the observation number}
    \item{year}{the year}
    \item{month}{the month}
    \item{rff}{the federal funds rate}
    \item{rt1}{the 1 year treasury bill rate}
    \item{rt3}{the 3 year treasury constant maturity rate}
    \item{rt10}{the 10 year treasury constant maturity rate}
  }
}
\references{
    Davidson, R.  and  James G.  MacKinnon (2004) \emph{Econometric Theory and Methods}, New York, Oxford University Press, chapter 13 and 14.
}
\keyword{datasets}