\name{Intratesq}
\docType{data}
\alias{Intratesq}
\title{Interest rates (quarterly)}
\description{
  quarterly observations from 1955-1 to 2001-4

 \emph{number of observations} :  188

 \emph{observation} :  Country

   \emph{country} :  United States

}
\usage{data(Intratesq)}
\format{A dataframe containing :
  \describe{
    \item{ID}{the observation number}
    \item{year}{the year}
    \item{quarter}{the quarter}
    \item{rff}{the federal funds rate}
    \item{rt1}{the 1 year treasury bill rate}
    \item{rt3}{the 3 year treasury constant maturity rate}
    \item{rt10}{the 10 year treasury constant maturity rate}
  }
}
\references{
    Davidson, R.  and  James G.  MacKinnon (2004) \emph{Econometric Theory and Methods}, New York, Oxford University Press, chapter 13.
}
\keyword{datasets}