\name{MW}
\docType{data}
\alias{MW}
\title{Growth of disposable income end treasury bill rate}
\description{
  quarterly observations from 1963-3 to 1975-4

 \emph{number of observations} :  50

 \emph{observation} :  Country

   \emph{country} :  United States

}
\usage{data(MW)}
\format{A dataframe containing :
  \describe{
    \item{ID}{contains the observation number}
    \item{x2}{the rate of growth of real U.S. disposable income, seasonally adjusted}
    \item{x3}{the U.S. treasury bill rate}
  }
}
\note{\dQuote{Both series have been normalized, as described in the paper.}}
\source{
    MacKinnon, James G.  and  Halbert  White (1985) \dQuote{Some heteroskedasticity consistent  covariance matrix estimators with improved finite sample properties}, \emph{Journal of Econometrics }, \bold{29}, 305-325.
}
\references{
    Davidson, R.  and  James G.  MacKinnon (2004) \emph{Econometric Theory and Methods}, New York, Oxford University Press, chapter 5.
}
\keyword{datasets}