\name{Partlin}
\docType{data}
\alias{Partlin}
\title{Artificial datas for estimating partially linear models}
\description{
  a cross-section

 \emph{number of observations} :  400

}
\usage{data(Partlin)}
\format{A dataframe containing :
  \describe{
    \item{x1}{artificial}
    \item{x2}{artificial}
    \item{x3}{artificial}
  }
}
\references{
    Davidson, R.  and  James G.  MacKinnon (2004) \emph{Econometric Theory and Methods}, New York, Oxford University Press, chapter 15.
}
\keyword{datasets}