\name{print.clos.etm}
\alias{print.clos.etm}
\title{
  Print function for 'clos.etm' objects
}
\description{
  Print method for object of class \code{clos.etm}
}
\usage{
\S3method{print}{clos.etm}(x, ...)
}
\arguments{
  \item{x}{An object of class \code{clos.etm}}
  \item{\dots}{Further arguments}
}
\value{
  No value returned
}
\author{
  Arthur Allignol, \email{arthur.allignol@gmail.com}
}
\seealso{
  \code{\link{clos}}
}
\keyword{print}