\name{print.etmCIF}
\alias{print.etmCIF}
\title{
  Print function for \code{cifETM} objects
}
\description{
    Print method for \code{cifETM} objects
}
\usage{
\S3method{print}{etmCIF}(x, ...)
}
\arguments{
  \item{x}{An object of class \code{etmCIF}}
  \item{\dots}{Further arguments}
}
\value{
  No value returned
}
\author{
  Arthur Allignol \email{arthur.allignol@gmail.com}
}
\seealso{
  \code{\link{etmCIF}}
}
\keyword{survival}
\keyword{print}
