% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETM.R
\name{as.matrix.ETM}
\alias{as.matrix.ETM}
\title{Get matrices out of an ETM object}
\usage{
\method{as.matrix}{ETM}(x, type = c("embedding", "beta"), which = c("topics", "words"), ...)
}
\arguments{
\item{x}{an object of class \code{ETM}}

\item{type}{character string with the type of information to extract: either 'beta', 'embedding'. Defaults to 'embedding'.}

\item{which}{if type is set to 'embedding', which embedding, either 'words' or 'topics'. Defaults to 'topics'.}

\item{...}{not used}
}
\description{
Convenience functions to extract embeddings of the cluster centers, the word embeddings
and the word emittance by each topic called beta which is technically the softmax-transformed inner product of word embedding and topic embeddings
}
