% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETM.R
\name{plot.ETM}
\alias{plot.ETM}
\title{Plot functionality for an ETM object}
\usage{
\method{plot}{ETM}(x, type = c("loss", "topics"), ...)
}
\arguments{
\item{x}{an object of class \code{ETM}}

\item{type}{character string with the type of plot, either 'loss' or 'topics'}

\item{...}{not used}
}
\description{
Convenience function allowing to plot the evolution of the loss
}
