% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETM.R
\name{predict.ETM}
\alias{predict.ETM}
\title{Predict to which ETM topic a text belongs}
\usage{
\method{predict}{ETM}(
  object,
  newdata,
  type = c("topics", "terms"),
  batch_size = nrow(newdata),
  normalize = TRUE,
  top_n = 10,
  ...
)
}
\arguments{
\item{object}{an object of class \code{ETM}}

\item{newdata}{bag of words document term matrix in \code{dgCMatrix} format}

\item{type}{either 'topics' or 'terms'}

\item{batch_size}{integer with the size of the batch}

\item{normalize}{logical indicating to normalize the bag of words data}

\item{top_n}{integer with number of most relevant words for each topic to extract}

\item{...}{not used}
}
\description{
Predict functionality for an \code{ETM} object
}
\examples{
\dontshow{if(require(torch) && torch::torch_is_installed())
\{
}

library(torch)
library(topicmodels.etm)
path  <- system.file(package = "topicmodels.etm", "example", "example_etm.ckpt")
model <- torch_load(path)

# Get most emitted words for each topic
terminology  <- predict(model, type = "terms", top_n = 5)
terminology

# Get topics probabilities for each document
path   <- system.file(package = "topicmodels.etm", "example", "example_dtm.rds")
dtm    <- readRDS(path)
dtm    <- head(dtm, n = 5)
scores <- predict(model, newdata = dtm, type = "topics")
scores

\dontshow{
\}
# End of main if statement running only if the torch is properly installed
}
}
\seealso{
\code{\link{ETM}}
}
