% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETM.R
\name{predict.ETM}
\alias{predict.ETM}
\title{Predict to which ETM topic a text belongs}
\usage{
\method{predict}{ETM}(
  object,
  newdata,
  type = c("topics", "terms"),
  batch_size = nrow(newdata),
  normalize = TRUE,
  top_n = 10,
  ...
)
}
\arguments{
\item{object}{an object of class \code{ETM}}

\item{newdata}{bag of words document term matrix in \code{dgCMatrix} format}

\item{type}{either 'topics' or 'terms'}

\item{batch_size}{integer with the size of the batch}

\item{normalize}{logical indicating to normalize the bag of words data}

\item{top_n}{integer with number of most relevant words for each topic to extract}

\item{...}{not used}
}
\description{
Predict functionality for an \code{ETM} object
}
