% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETM.R
\name{summary.ETM}
\alias{summary.ETM}
\title{Project ETM embeddings using UMAP}
\usage{
\method{summary}{ETM}(object, type = c("umap"), n_components = 2, top_n = 20, ...)
}
\arguments{
\item{object}{object of class \code{ETM}}

\item{type}{character string with the type of summary. Defaults to 'umap'.}

\item{n_components}{the dimension of the space to embed into. Passed on to \code{\link[uwot]{umap}}}

\item{top_n}{passed on to \code{\link{predict.ETM}} to get the top_n most relevant words for each topic in the 2-dimensional space}

\item{...}{further arguments passed onto \code{\link[uwot]{umap}}}
}
\description{
Uses the uwot package to map the word embeddings and the center of the topic embeddings to a 2-dimensional space
}
\seealso{
\code{\link[uwot]{umap}}
}
